fix: connection update to multiple databases with a single client
