+++ .env.example
@@ -22,20 +22,6 @@ MONGO_SERVER_SELECTION_TIMEOUT_MS=2000
 MONGO_WAITING_QUEUE_MS=2000
 MONGO_HEARTBEAT_FREQUENCY_MS=5000
 
-ECOMMERCE_HISTORY_MONGO_HOST="pagopa-ecommerce-history-helpdesk-mongo"
-ECOMMERCE_HISTORY_MONGO_PORT=27017
-ECOMMERCE_HISTORY_MONGO_USERNAME=admin
-ECOMMERCE_HISTORY_MONGO_PASSWORD=password
-ECOMMERCE_HISTORY_MONGO_SSL_ENABLED=false
-ECOMMERCE_HISTORY_MONGO_MIN_POOL_SIZE=0
-ECOMMERCE_HISTORY_MONGO_MAX_POOL_SIZE=20
-ECOMMERCE_HISTORY_MONGO_MAX_IDLE_TIMEOUT_MS=60000
-ECOMMERCE_HISTORY_MONGO_CONNECTION_TIMEOUT_MS=1000
-ECOMMERCE_HISTORY_MONGO_SOCKET_TIMEOUT_MS=10000
-ECOMMERCE_HISTORY_MONGO_SERVER_SELECTION_TIMEOUT_MS=2000
-ECOMMERCE_HISTORY_MONGO_WAITING_QUEUE_MS=2000
-ECOMMERCE_HISTORY_MONGO_HEARTBEAT_FREQUENCY_MS=5000
-
 SEARCH_DEAD_LETTER_QUEUE_MAPPING={ ALL: '*', ECOMMERCE: 'pagopa-ecommerce-transactions-dead-letter-queue', NOTIFICATIONS_SERVICE: 'pagopa-ecommerce-notifications-service-errors-queue' }
 SEARCH_PM_IN_ECOMMERCE_HISTORY_ENABLED=false
