+++ docker-compose.yml
@@ -56,31 +56,6 @@ services:
     depends_on:
       - mongo
 
-  mongo-ecommerce-history:
-    container_name: pagopa-ecommerce-history-helpdesk-mongo
-    image: mongo
-    env_file: docker/mongoDB/conf/mongodb-ecommerce-history.env
-    ports:
-      - "27018:27017"
-    networks:
-      - pagopa-ecommerce-net
-    healthcheck:
-      test: echo "db.stats().ok" | mongo --quiet
-      interval: 10s
-    volumes:
-      - ./docker/mongoDB/mongo-ecommerce-history-init.js:/docker-entrypoint-initdb.d/mongo-init.js:ro
-
-  mongo-express-ecommerce-history:
-    container_name: pagopa-ecommerce-history-helpdesk-express
-    image: mongo-express
-    env_file: docker/mongoDB/conf/mongo-express-ecommerce-history.env
-    ports:
-      - "8082:8081"
-    networks:
-      - pagopa-ecommerce-net
-    depends_on:
-      - mongo-ecommerce-history
-
   pagopa-pdv-mock:
     container_name: pagopa-pdv-mock-helpdesk
     build:
