--- docker/mongoDB/conf/mongo-express-ecommerce-history.env
@@ -1,3 +1,0 @@
-MONGO_INITDB_ROOT_USERNAME=admin
-MONGO_INITDB_ROOT_PASSWORD=password
-ME_CONFIG_MONGODB_URL=mongodb://$MONGO_INITDB_ROOT_USERNAME:$MONGO_INITDB_ROOT_PASSWORD@pagopa-ecommerce-history-helpdesk-mongo:27017/
