--- docker/mongoDB/conf/mongodb-ecommerce-history.env
@@ -1,3 +1,0 @@
-MONGO_INITDB_ROOT_USERNAME=admin
-MONGO_INITDB_ROOT_PASSWORD=password
-MONGO_INITDB_DATABASE=ecommerce-history
