--- src/main/kotlin/it/pagopa/ecommerce/helpdesk/configurations/MongoEcommerceHistoryConfig.kt
@@ -1,31 +1,0 @@
-package it.pagopa.ecommerce.helpdesk.configurations
-
-import com.mongodb.ConnectionString
-import com.mongodb.MongoClientSettings
-import com.mongodb.reactivestreams.client.MongoClient
-import com.mongodb.reactivestreams.client.MongoClients
-import org.springframework.beans.factory.annotation.Qualifier
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.mongodb.core.ReactiveMongoTemplate
-
-@Configuration
-class MongoEcommerceHistoryConfig {
-    @Bean(name = ["ecommerceHistoryReactiveMongoClient"])
-    fun ecommerceHistoryReactiveMongoClient(
-        @Value("\${ecommerce_history.mongodb.uri}") uri: String,
-    ): MongoClient {
-        val connectionString = ConnectionString(uri)
-        val settings = MongoClientSettings.builder().applyConnectionString(connectionString).build()
-        return MongoClients.create(settings)
-    }
-
-    @Bean(name = ["ecommerceHistoryReactiveMongoTemplate"])
-    fun ecommerceHistoryReactiveMongoTemplate(
-        @Qualifier("ecommerceHistoryReactiveMongoClient") mongoClient: MongoClient,
-        @Value("\${ecommerce_history.mongodb.database}") database: String
-    ): ReactiveMongoTemplate {
-        return ReactiveMongoTemplate(mongoClient, database)
-    }
-}
