+++ src/main/resources/application.properties
@@ -27,13 +27,11 @@ pm.oracle.password=\${PM_ORACLE_PASSWORD}
 #spring.data.mongodb.uri=mongodb://\${MONGO_USERNAME}:\${MONGO_PASSWORD}@\${MONGO_HOST}:\${MONGO_PORT}/?ssl=\${MONGO_SSL_ENABLED}&readPreference=secondary&minPoolSize=\${MONGO_MIN_POOL_SIZE}&maxPoolSize=\${MONGO_MAX_POOL_SIZE}&maxIdleTimeMS=\${MONGO_MAX_IDLE_TIMEOUT_MS}&connectTimeoutMS=\${MONGO_CONNECTION_TIMEOUT_MS}&socketTimeoutMS=\${MONGO_SOCKET_TIMEOUT_MS}&serverSelectionTimeoutMS=\${MONGO_SERVER_SELECTION_TIMEOUT_MS}&waitQueueTimeoutMS=\${MONGO_WAITING_QUEUE_MS}&heartbeatFrequencyMS=\${MONGO_HEARTBEAT_FREQUENCY_MS}\${MONGO_REPLICA_SET_OPTION:}
 #spring.data.mongodb.database=ecommerce
 
-#Ecommerce mongo database configuration
-ecommerce.mongodb.uri=mongodb://\${MONGO_USERNAME}:\${MONGO_PASSWORD}@\${MONGO_HOST}:\${MONGO_PORT}/?ssl=\${MONGO_SSL_ENABLED}&readPreference=secondary&minPoolSize=\${MONGO_MIN_POOL_SIZE}&maxPoolSize=\${MONGO_MAX_POOL_SIZE}&maxIdleTimeMS=\${MONGO_MAX_IDLE_TIMEOUT_MS}&connectTimeoutMS=\${MONGO_CONNECTION_TIMEOUT_MS}&socketTimeoutMS=\${MONGO_SOCKET_TIMEOUT_MS}&serverSelectionTimeoutMS=\${MONGO_SERVER_SELECTION_TIMEOUT_MS}&waitQueueTimeoutMS=\${MONGO_WAITING_QUEUE_MS}&heartbeatFrequencyMS=\${MONGO_HEARTBEAT_FREQUENCY_MS}\${MONGO_REPLICA_SET_OPTION:}
-ecommerce.mongodb.database=ecommerce
+#Ecommerce mongo database configurations
+mongodb.uri=mongodb://\${MONGO_USERNAME}:\${MONGO_PASSWORD}@\${MONGO_HOST}:\${MONGO_PORT}/?ssl=\${MONGO_SSL_ENABLED}&readPreference=secondary&minPoolSize=\${MONGO_MIN_POOL_SIZE}&maxPoolSize=\${MONGO_MAX_POOL_SIZE}&maxIdleTimeMS=\${MONGO_MAX_IDLE_TIMEOUT_MS}&connectTimeoutMS=\${MONGO_CONNECTION_TIMEOUT_MS}&socketTimeoutMS=\${MONGO_SOCKET_TIMEOUT_MS}&serverSelectionTimeoutMS=\${MONGO_SERVER_SELECTION_TIMEOUT_MS}&waitQueueTimeoutMS=\${MONGO_WAITING_QUEUE_MS}&heartbeatFrequencyMS=\${MONGO_HEARTBEAT_FREQUENCY_MS}\${MONGO_REPLICA_SET_OPTION:}
+mongodb.ecommerce.database=ecommerce
+mongodb.ecommerce_history.database=ecommerce-history
 
-#Ecommerce History mongo database configuration
-ecommerce_history.mongodb.uri=mongodb://\${ECOMMERCE_HISTORY_MONGO_USERNAME}:\${ECOMMERCE_HISTORY_MONGO_PASSWORD}@\${ECOMMERCE_HISTORY_MONGO_HOST}:\${ECOMMERCE_HISTORY_MONGO_PORT}/?ssl=\${ECOMMERCE_HISTORY_MONGO_SSL_ENABLED}&readPreference=secondary&minPoolSize=\${ECOMMERCE_HISTORY_MONGO_MIN_POOL_SIZE}&maxPoolSize=\${ECOMMERCE_HISTORY_MONGO_MAX_POOL_SIZE}&maxIdleTimeMS=\${ECOMMERCE_HISTORY_MONGO_MAX_IDLE_TIMEOUT_MS}&connectTimeoutMS=\${ECOMMERCE_HISTORY_MONGO_CONNECTION_TIMEOUT_MS}&socketTimeoutMS=\${ECOMMERCE_HISTORY_MONGO_SOCKET_TIMEOUT_MS}&serverSelectionTimeoutMS=\${ECOMMERCE_HISTORY_MONGO_SERVER_SELECTION_TIMEOUT_MS}&waitQueueTimeoutMS=\${ECOMMERCE_HISTORY_MONGO_WAITING_QUEUE_MS}&heartbeatFrequencyMS=\${ECOMMERCE_HISTORY_MONGO_HEARTBEAT_FREQUENCY_MS}\${ECOMMERCE_HISTORY_MONGO_REPLICA_SET_OPTION:}
-ecommerce_history.mongodb.database=ecommerce-history
 
 deadLetter.queueMapping=\${SEARCH_DEAD_LETTER_QUEUE_MAPPING}
 #PDV configuration
