+++ src/test/kotlin/it/pagopa/ecommerce/helpdesk/configurations/MongoEcommerceConfigTest.kt
@@ -9,7 +9,8 @@ class MongoEcommerceConfigTest {
     private val mongoEcommerceConfig = MongoEcommerceConfig()
     private val uri = "mongodb://localhost:27017/test"
     private val mongoClient: MongoClient = mongoEcommerceConfig.ecommerceReactiveMongoClient(uri)
-    val database = "ecommerce"
+    val database_ecommerce = "ecommerce"
+    val database_ecommerce_history = "ecommerce-history"
 
     @Test
     fun `ecommerceReactiveMongoClient should create MongoClient with correct settings`() {
@@ -19,10 +20,23 @@ class MongoEcommerceConfigTest {
     @Test
     fun `ecommerceReactiveMongoTemplate should create ReactiveMongoTemplate with correct client and database`() {
         val reactiveMongoTemplate =
-            mongoEcommerceConfig.ecommerceReactiveMongoTemplate(mongoClient, database)
+            mongoEcommerceConfig.ecommerceReactiveMongoTemplate(mongoClient, database_ecommerce)
 
         assertThat(reactiveMongoTemplate).isNotNull
         assertThat(reactiveMongoTemplate.mongoDatabaseFactory.mongoDatabase.block()?.name)
-            .isEqualTo(database)
+            .isEqualTo(database_ecommerce)
+    }
+
+    @Test
+    fun `ecommerceHistoryReactiveMongoTemplate should create ReactiveMongoTemplate with correct client and database`() {
+        val reactiveMongoTemplate =
+            mongoEcommerceConfig.ecommerceHistoryReactiveMongoTemplate(
+                mongoClient,
+                database_ecommerce_history
+            )
+
+        assertThat(reactiveMongoTemplate).isNotNull
+        assertThat(reactiveMongoTemplate.mongoDatabaseFactory.mongoDatabase.block()?.name)
+            .isEqualTo(database_ecommerce_history)
     }
 }
