+++ src/test/resources/application.properties
@@ -26,13 +26,10 @@ pm.oracle.password=test
 #spring.data.mongodb.uri=mongodb://mongo:mongo@$mongo:6080/?ssl=true&replicaSet=globaldb
 #spring.data.mongodb.database=ecommerce
 
-#Ecommerce mongo database configuration
-ecommerce.mongodb.uri=mongodb://mongo:mongo@$mongo:6080/?ssl=true&replicaSet=globaldb
-ecommerce.mongodb.database=ecommerce
-
-#Ecommerce History mongo database configuration
-ecommerce_history.mongodb.uri=mongodb://mongo:mongo@$mongo:6081/?ssl=true&replicaSet=globaldb
-ecommerce_history.mongodb.database=ecommerce-history
+#Ecommerce mongo database configurations
+mongodb.uri=mongodb://mongo:mongo@$mongo:6081/?ssl=true&replicaSet=globaldb
+mongodb.ecommerce.database=ecommerce
+mongodb.ecommerce_history.database=ecommerce-history
 
 
 deadLetter.queueMapping={ ALL: '*', ECOMMERCE: 'pagopa-ecommerce-transactions-dead-letter-queue', NOTIFICATIONS_SERVICE: 'pagopa-ecommerce-notifications-service-errors-queue' }
